+++ src/document_buffer.cpp
@@ -72,7 +72,7 @@ void document_buffer::clear() {
 	markers.clear();
 }
 
-int document_buffer::next_marker_index(long position, marker_type type) const noexcept {
+int document_buffer::next_marker_index(int position, marker_type type) const noexcept {
 	for (size_t i = 0; i < markers.size(); ++i) {
 		if (static_cast<long>(markers[i].pos) > position && markers[i].type == type) {
 			return static_cast<int>(i);
@@ -81,7 +81,7 @@ int document_buffer::next_marker_index(long position, marker_type type) const no
 	return -1;
 }
 
-int document_buffer::previous_marker_index(size_t position, marker_type type) const noexcept {
+int document_buffer::previous_marker_index(int position, marker_type type) const noexcept {
 	int current_index = -1;
 	for (size_t i = 0; i < markers.size(); ++i) {
 		if (markers[i].pos >= position && markers[i].type == type) {
@@ -114,7 +114,7 @@ int document_buffer::current_marker_index(size_t position, marker_type type) con
 	return -1;
 }
 
-int document_buffer::next_heading_marker_index(long position, int level) const {
+int document_buffer::next_heading_marker_index(int position, int level) const {
 	const auto heading_markers = get_heading_markers();
 	for (size_t i = 0; i < heading_markers.size(); ++i) {
 		if (static_cast<long>(heading_markers[i]->pos) > position) {
@@ -126,7 +126,7 @@ int document_buffer::next_heading_marker_index(long position, int level) const {
 	return -1;
 }
 
-int document_buffer::previous_heading_marker_index(size_t position, int level) const {
+int document_buffer::previous_heading_marker_index(int position, int level) const {
 	const auto heading_markers = get_heading_markers();
 	int current_index = -1;
 	for (size_t i = 0; i < heading_markers.size(); ++i) {
